Move GlobalPrideSettingComponent class into its own file
